+++ mojo/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.2.12"
+__version__ = "1.2.13"
 
 from mojo.helpers.response import JsonResponse